Revert "remove `start` method from sdk"
This reverts commit 15819fe1bb973834d733f21ebd8909bea4e37931.
