+++ example/index.js
@@ -1,5 +1,4 @@
-const http = require("http");
-const { flow, act, check } = require("../dist"); // or "automated-qa-sdk"
+const { flow, act, check, start } = require("../dist"); // or "automated-qa-sdk"
 let request = require("request-promise-native");
 
 request = request.defaults({
@@ -145,42 +144,4 @@ flow("Basic API functionality", () => {
   check("first key should be invalid");
 });
 
-function start() {
-  const env = process.env.NODE_ENV || "development";
-  const port = process.env.PORT || "3000";
-  const server = http.createServer(async (req, res) => {
-    if (env !== "test") {
-      console.log("%s %s", req.method, req.url);
-    }
-
-    // Hande 404s
-    if (req.url !== "/") {
-      return renderJson(res, 404, error4002);
-    }
-
-    // Handle page (GET /)
-    if (req.method !== "POST") {
-      return renderJson(res, 200, {});
-    }
-
-    // Run flows
-    try {
-      const flows = await run();
-      renderJson(
-        res,
-        200,
-        flows.map((f) => ({
-          assertions: f.assertions,
-          name: f.name,
-        }))
-      );
-    } catch (e) {
-      log("error", e);
-      return renderJson(res, 500, error5000);
-    }
-  });
-
-  server.listen(parseInt(port, 10));
-  console.log("automated-qa worker listening on port %s", port);
-}
 start();
